+++ src/main/java/org/opensearch/security/privileges/SystemIndexAccessEvaluator.java
@@ -49,7 +49,7 @@ import org.opensearch.security.support.WildcardMatcher;
 import org.opensearch.security.user.User;
 import org.opensearch.tasks.Task;
 
-import static org.opensearch.security.privileges.PrivilegesEvaluator.isClusterPerm;
+import static org.opensearch.security.privileges.PrivilegesEvaluatorImpl.isClusterPerm;
 
 /**
  * This class performs authorization on requests targeting system indices