+++ src/test/java/org/opensearch/security/resources/ResourceAccessHandlerTest.java
@@ -23,7 +23,6 @@ import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.security.auth.UserSubjectImpl;
 import org.opensearch.security.configuration.AdminDNs;
-import org.opensearch.security.privileges.PrivilegesEvaluator;
 import org.opensearch.security.resources.sharing.ResourceSharing;
 import org.opensearch.security.resources.sharing.ShareWith;
 import org.opensearch.security.securityconf.FlattenedActionGroups;
@@ -51,8 +50,6 @@ public class ResourceAccessHandlerTest {
     private ResourceSharingIndexHandler sharingIndexHandler;
     @Mock
     private AdminDNs adminDNs;
-    @Mock
-    private PrivilegesEvaluator privilegesEvaluator;
 
     @Mock
     private ResourcePluginInfo resourcePluginInfo;
@@ -69,7 +66,7 @@ public class ResourceAccessHandlerTest {
     public void setup() {
         threadContext = new ThreadContext(Settings.EMPTY);
         when(threadPool.getThreadContext()).thenReturn(threadContext);
-        handler = new ResourceAccessHandler(threadPool, sharingIndexHandler, adminDNs, privilegesEvaluator, resourcePluginInfo);
+        handler = new ResourceAccessHandler(threadPool, sharingIndexHandler, adminDNs, resourcePluginInfo);
 
         // For tests that verify permission with action-group
         when(resourcePluginInfo.flattenedForType(any())).thenReturn(mock(FlattenedActionGroups.class));